+++ src/otx/engine/utils/hpo/__init__.py
@@ -1,8 +1,9 @@
 # Copyright (C) 2024 Intel Corporation
 # SPDX-License-Identifier: Apache-2.0
 #
-"""Files to run HPO in the OTX Engine."""
+"""Functions and Classes to run HPO in the engine."""
 
-from .hpo import execute_hpo, update_hyper_parameter
+from .hpo import execute_hpo
+from .hpo_trial import update_hyper_parameter
 
 __all__ = ["execute_hpo", "update_hyper_parameter"]
